+++ app/src/main/java/ru/ds/gitapp/ui/gitrepo/ReposViewModelFactory.kt
@@ -1,12 +1,12 @@
-package ru.ds.gitapp.ui.users
+package ru.ds.gitapp.ui.gitrepo
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import ru.ds.gitapp.data.remote.GitUserRep
 
 
-class ReposUserViewModelFactory(private val repo: GitUserRep) : ViewModelProvider.Factory {
+class ReposViewModelFactory(private val repo: GitUserRep) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return UserRepositoryViewModel(repo) as T
+        return GITRepositoryViewModel(repo) as T
     }
 }
