+++ app/src/main/java/ru/ds/gitapp/ui/gitusers/ReposUserViewModelFactory.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.ViewModelProvider
 import ru.ds.gitapp.data.remote.GitUserRep
 
 
-class ReposViewModelFactory(private val repo: GitUserRep) : ViewModelProvider.Factory {
+class ReposUserViewModelFactory(private val repo: GitUserRep) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return GITRepositoryViewModel(repo) as T
+        return UserRepositoryViewModel(repo) as T
     }
 }
