+++ app/src/main/java/ru/ds/gitapp/ui/gitusers/UserRepositoryViewModel.kt
@@ -8,7 +8,7 @@ import io.reactivex.rxjava3.kotlin.subscribeBy
 import ru.ds.gitapp.data.remote.GitUserEntity
 import ru.ds.gitapp.data.remote.GitUserRep
 
-class GITRepositoryViewModel(private val gitUserRep: GitUserRep) : ViewModel() {
+class UserRepositoryViewModel(private val gitUserRep: GitUserRep) : ViewModel() {
     private val _repo = MutableLiveData<List<GitUserEntity>>()
     val repo: LiveData<List<GitUserEntity>> = _repo
 
@@ -19,19 +19,17 @@ class GITRepositoryViewModel(private val gitUserRep: GitUserRep) : ViewModel() {
     private val compositeDisposable: CompositeDisposable = CompositeDisposable()
 
     //метод - подписка на обновления с сервера
-    fun onShowRepository(username: String) {
+    fun onShowUserRepository(username: String) {
         _inProgerss.postValue(true) // устанавливаем progress Bar
         compositeDisposable
             .add(gitUserRep
-                .getUsers(username)
+                .getUsersRep(username)
                 .subscribeBy {
                     _inProgerss.postValue(false) // убираем progress Bar
                     _repo.postValue(it)// как только приходит результат отправляем его
                 }
             )
     }
-
-
 
     override fun onCleared() {
         compositeDisposable.clear()
